requirements.txt
Flask==2.0.3
openai==0.11.3
python-dotenv==0.19.2
markdown2==2.4.0
selenium==4.1.0
pdfkit==0.6.1
